Allow max and min alter settings
This was a particularly nasty refactoring. The study now has two alter
counting parameters: min and max. The interview structure was previously
fixed when the interview started (number of questions, number of
expected responses, etc). The interview structure is now recalculated
every time a new number of alters is set.